pyproject.toml
[tool.poetry]
name = "budget-bot-2"
version = "0.1.0"
description = ""
authors = ["Pavel Maksimov <maksimov.pavel.an@yandex.ru>"]
readme = "README.md"
packages = [{include = "budget_bot_2", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.13"}
alembic = "^1.11.0"
asyncpg = "^0.27.0"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.3"
matplotlib = "^3.7.1"
jinja2 = "^3.1.2"
pillow = "^9.5.0"
pdfkit = "^1.0.0"
wkhtmltopdf = "^0.2"
psycopg2 = "^2.9.6"
coverage = "^7.2.7"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pylint = "^2.17.4"
pre-commit = "^3.3.2"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Configuration for various linters, formatters, checkers etc.
[tool.black]
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
filter_files = true

[tool.pylint.messages_control]
score=false
reports=false
ignore-paths = ["alembic"]
extension-pkg-whitelist = "pydantic"
disable = [
  "missing-docstring",
  "too-few-public-methods",
  "unused-argument",
  "not-callable",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
